Refactor : 프론트 의견 반영 search POST 매핑 수정 - search 용 dto 생성
