+++ src/main/java/com/unity/goods/domain/goods/controller/GoodsController.java
@@ -7,8 +7,8 @@ import com.unity.goods.domain.goods.dto.UpdateGoodsStateDto.UpdateGoodsStateRequ
 import com.unity.goods.domain.goods.dto.UploadGoodsDto;
 import com.unity.goods.domain.goods.service.GoodsService;
 import com.unity.goods.global.jwt.UserDetailsImpl;
-import com.unity.goods.infra.document.GoodsDocument;
-import com.unity.goods.infra.dto.SearchedGoods;
+import com.unity.goods.infra.dto.SearchDto.SearchRequest;
+import com.unity.goods.infra.dto.SearchDto.SearchedGoods;
 import com.unity.goods.infra.service.GoodsSearchService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -92,11 +92,12 @@ public class GoodsController {
     goodsService.deleteGoods(member, goodsId);
     return ResponseEntity.ok().build();
   }
-  
-  @GetMapping("/search")
+
+  @PostMapping("/search")
   public ResponseEntity<?> search(
-      @RequestParam(name = "keyword") String keyword,
+      @RequestBody SearchRequest searchRequest,
       @PageableDefault Pageable pageable) {
+    String keyword = searchRequest.getKeyword();
     Page<SearchedGoods> goodsDocumentPage = goodsSearchService.search(keyword, pageable);
     return ResponseEntity.ok(goodsDocumentPage);
   }
